Rename: user-stickers.entity로 파일명 변경 #112
