+++ src/domain/users/entities/user-stickers.entity.ts
